LayoutTests/http/tests/security/sandboxed-iframe-DENIED-modals-expected.txt

CONSOLE MESSAGE: Use of window.alert is not allowed in a sandboxed frame when the allow-modals flag is not set.
CONSOLE MESSAGE: window.alert returned undefined
CONSOLE MESSAGE: Use of window.confirm is not allowed in a sandboxed frame when the allow-modals flag is not set.
CONSOLE MESSAGE: window.confirm returned false
CONSOLE MESSAGE: Use of window.prompt is not allowed in a sandboxed frame when the allow-modals flag is not set.
CONSOLE MESSAGE: window.prompt returned null
CONSOLE MESSAGE: window.print returned undefined
CONSOLE MESSAGE: Use of window.print is not allowed in a sandboxed frame when the allow-modals flag is not set.
CONSOLE MESSAGE: Use of window.print is not allowed in a sandboxed frame when the allow-modals flag is not set.
CONSOLE MESSAGE: document.execCommand returned false
This test passes if opening modal dialogs is blocked and the corresponding calls return immediately with the expected values.
